+++ backend/src/neuroagent/tools/entitycore_brainregion_getall.py
@@ -1,6 +1,6 @@
 """Get Brain Region tool."""
 
-from typing import ClassVar, Literal
+from typing import ClassVar
 
 from httpx import AsyncClient
 from pydantic import Field
@@ -32,12 +32,6 @@ class BrainRegionGetAllInput(
         title="Brain Region Name Search",
         description="Perform semantic search to find brain regions by their names. Enter any text related to a brain region name (e.g., 'hippocampus', 'frontal cortex', 'amygdala') and receive results ranked by semantic similarity to your query.",
     )
-    hierarchy_id: Literal[
-        "e3e70682-c209-4cac-a29f-6fbed82c07cd", "e3fdfcc0-6807-4be1-aefc-b3f9116f6ced"
-    ] = Field(
-        default="e3e70682-c209-4cac-a29f-6fbed82c07cd",
-        description="The hierarchy ID for brain regions. The default value is the most commonly used hierarchy ID called 'aibs'. The second one is : `Julich-Brain Probabilistic Cytoarchitectonic Atlas`.",
-    )  # type: ignore
 
 
 class BrainRegionGetAllTool(BaseTool):
@@ -53,6 +47,8 @@ class BrainRegionGetAllTool(BaseTool):
         "Find an electrical recording in the thalamus and make an in depth analysis of it.",
         "Are there circuits in the sscx ?",
         "Find all of the thalamical neurons.",
+        "Show me morphologies in the rat's isocortex.",
+        "Find a circuit from the human's hippocampus.",
     ]
     description: ClassVar[
         str
@@ -87,7 +83,6 @@ class BrainRegionGetAllTool(BaseTool):
         """
         query_params = self.input_schema.model_dump(exclude_defaults=True, mode="json")
         query_params["page_size"] = self.input_schema.page_size
-        query_params["hierarchy_id"] = self.input_schema.hierarchy_id
 
         headers: dict[str, str] = {}
         if self.metadata.vlab_id is not None: