+++ backend/src/neuroagent/tools/entitycore_simulationexecution_getall.py
@@ -1,7 +1,6 @@
 """Get All SimulationExecution tool."""
 
 from typing import ClassVar
-from uuid import UUID
 
 from httpx import AsyncClient
 from pydantic import Field
@@ -16,14 +15,6 @@ from neuroagent.tools.base_tool import BaseTool, EntitycoreMetadata
 class SimulationExecutionGetAllInput(ReadManySimulationExecutionGetParametersQuery):
     """Inputs for the simulation-execution get all tool."""
 
-    within_brain_region_hierarchy_id: UUID | None = Field(
-        default=UUID("e3e70682-c209-4cac-a29f-6fbed82c07cd"),
-        description="The hierarchy ID for brain regions. The default value is the most commonly used hierarchy ID.",
-    )
-    within_brain_region_brain_region_id: UUID | None = Field(
-        default=None,
-        description="ID of the brain region of interest in UUID format.",
-    )
     page_size: int = Field(
         ge=1,
         le=10,